Fixes #1 , an issue with wrong URLs for the notifications.
The simple fix should be just replacing 'pulls' with 'pull' in
the URL, but it doesn't seem like a sustainable solution.
GitHub's Notifications API doesn't return the 'html_url' for the
links, so the right way to get it is another API call. This API call
is done by the time the link is clicked, making sure that it won't
go over the rate limit.